+++ clang/lib/Sema/SemaTemplateInstantiateDecl.cpp
@@ -5015,11 +5015,8 @@ void Sema::InstantiateFunctionDefinition(SourceLocation PointOfInstantiation,
     llvm::raw_string_ostream OS(M.Detail);
     Function->getNameForDiagnostic(OS, getPrintingPolicy(),
                                    /*Qualified=*/true);
-    if (llvm::isTimeTraceVerbose()) {
-      auto Loc = SourceMgr.getExpansionLoc(Function->getLocation());
-      M.File = SourceMgr.getFilename(Loc);
-      M.Line = SourceMgr.getExpansionLineNumber(Loc);
-    }
+    if (llvm::isTimeTraceVerbose())
+      SourceMgr.setLocationForTimeTrace(Function->getLocation(), M);
     return M;
   });